+++ vscode/test/e2e/command-custom.test.ts
@@ -2,7 +2,12 @@ import { expect } from '@playwright/test'
 import * as mockServer from '../fixtures/mock-server'
 
 import { sidebarExplorer, sidebarSignin } from './common'
-import { type DotcomUrlOverride, test as baseTest, withPlatformSlashes, assertEvents } from './helpers'
+import {
+    type DotcomUrlOverride,
+    test as baseTest,
+    withPlatformSlashes,
+    type ExpectedEvents,
+} from './helpers'
 
 const test = baseTest.extend<DotcomUrlOverride>({ dotcomUrl: mockServer.SERVER_URL })
 
@@ -10,7 +15,20 @@ test.beforeEach(() => {
     mockServer.resetLoggedEvents()
 })
 
-test('create a new user command via the custom commands menu', async ({ page, sidebar }) => {
+test.extend<ExpectedEvents>({
+    // list of events we expect this test to log, add to this list as needed
+    expectedEvents: [
+        'CodyVSCodeExtension:auth:clickOtherSignInOptions',
+        'CodyVSCodeExtension:login:clicked',
+        'CodyVSCodeExtension:auth:selectSigninMenu',
+        'CodyVSCodeExtension:auth:fromToken',
+        'CodyVSCodeExtension:Auth:connected',
+        'CodyVSCodeExtension:menu:custom:build:clicked',
+        'CodyVSCodeExtension:chat-question:submitted',
+        'CodyVSCodeExtension:chat-question:executed',
+        'CodyVSCodeExtension:command:custom:executed',
+    ],
+})('create a new user command via the custom commands menu', async ({ page, sidebar }) => {
     // Sign into Cody
     await sidebarSignin(page, sidebar)
 
@@ -85,19 +103,24 @@ test('create a new user command via the custom commands menu', async ({ page, si
     // Confirm the command prompt is displayed in the chat panel on execution
     const chatPanel = page.frameLocator('iframe.webview').last().frameLocator('iframe')
     await expect(chatPanel.getByText(prompt)).toBeVisible()
-
-    const expectedEvents = [
-        'CodyVSCodeExtension:menu:custom:build:clicked',
-        'CodyVSCodeExtension:chat-question:submitted',
-        'CodyVSCodeExtension:chat-question:executed',
-        'CodyVSCodeExtension:command:custom:executed',
-    ]
-    await assertEvents(mockServer.loggedEvents, expectedEvents)
 })
 
 // NOTE: If no custom commands are showing up in the command menu, it might
 // indicate a breaking change during the custom command building step.
-test('execute custom commands with context defined in cody.json', async ({ page, sidebar }) => {
+test.extend<ExpectedEvents>({
+    // list of events we expect this test to log, add to this list as needed
+    expectedEvents: [
+        'CodyVSCodeExtension:auth:clickOtherSignInOptions',
+        'CodyVSCodeExtension:login:clicked',
+        'CodyVSCodeExtension:auth:selectSigninMenu',
+        'CodyVSCodeExtension:auth:fromToken',
+        'CodyVSCodeExtension:Auth:connected',
+        'CodyVSCodeExtension:menu:command:custom:clicked',
+        'CodyVSCodeExtension:command:custom:executed',
+        'CodyVSCodeExtension:chat-question:submitted',
+        'CodyVSCodeExtension:chat-question:executed',
+    ],
+})('execute custom commands with context defined in cody.json', async ({ page, sidebar }) => {
     // Sign into Cody
     await sidebarSignin(page, sidebar)
 
@@ -130,13 +153,15 @@ test('execute custom commands with context defined in cody.json', async ({ page,
 
     await expect(chatPanel.getByText('Add four context files from the current directory.')).toBeVisible()
     // Show the current file numbers used as context
-    await expect(chatPanel.getByText('✨ Context: 66 lines from 5 files')).toBeVisible()
-    await chatPanel.getByText('✨ Context: 66 lines from 5 files').click()
+    await expect(chatPanel.getByText('✨ Context: 61 lines from 5 files')).toBeVisible()
+    await chatPanel.getByText('✨ Context: 61 lines from 5 files').click()
     // Display the context files to confirm no hidden files are included
-    await expect(chatPanel.locator('span').filter({ hasText: '@Main.java:1-9' })).toBeVisible()
-    await expect(chatPanel.locator('span').filter({ hasText: '@buzz.test.ts:1-12' })).toBeVisible()
-    await expect(chatPanel.locator('span').filter({ hasText: '@buzz.ts:1-15' })).toBeVisible()
-    await expect(chatPanel.locator('span').filter({ hasText: '@index.html:1-11' })).toBeVisible()
+    await expect(chatPanel.locator('span').filter({ hasText: '@.mydotfile:1-2' })).not.toBeVisible()
+    await expect(chatPanel.locator('span').filter({ hasText: '@error.ts:1-10' })).toBeVisible()
+    await expect(chatPanel.locator('span').filter({ hasText: '@Main.java:1-10' })).toBeVisible()
+    await expect(chatPanel.locator('span').filter({ hasText: '@buzz.test.ts:1-13' })).toBeVisible()
+    await expect(chatPanel.locator('span').filter({ hasText: '@buzz.ts:1-16' })).toBeVisible()
+    await expect(chatPanel.locator('span').filter({ hasText: '@index.html:1-12' })).toBeVisible()
 
     /* Test: context.filePath with filePath command */
 
@@ -157,14 +182,14 @@ test('execute custom commands with context defined in cody.json', async ({ page,
     await page.getByPlaceholder('Search command to run...').fill('directory')
     await page.keyboard.press('Enter')
     await expect(chatPanel.getByText('Directory has one context file.')).toBeVisible()
-    await expect(chatPanel.getByText('✨ Context: 15 lines from 2 file')).toBeVisible()
-    await chatPanel.getByText('✨ Context: 15 lines from 2 file').click()
+    await expect(chatPanel.getByText('✨ Context: 14 lines from 2 file')).toBeVisible()
+    await chatPanel.getByText('✨ Context: 14 lines from 2 file').click()
     await expect(
-        chatPanel.locator('span').filter({ hasText: withPlatformSlashes('@lib/batches/env/var.go:1-1') })
+        chatPanel.locator('span').filter({ hasText: withPlatformSlashes('@lib/batches/env/var.go:1-2') })
     ).toBeVisible()
     // Click on the file link should open the 'var.go file in the editor
     await chatPanel
-        .getByRole('button', { name: withPlatformSlashes('@lib/batches/env/var.go:1-1') })
+        .getByRole('button', { name: withPlatformSlashes('@lib/batches/env/var.go:1-2') })
         .click()
     await expect(page.getByRole('tab', { name: 'var.go' })).toBeVisible()
 
@@ -179,21 +204,24 @@ test('execute custom commands with context defined in cody.json', async ({ page,
     // The files from the open tabs should be added as context
     await expect(chatPanel.getByText('✨ Context: 14 lines from 2 files')).toBeVisible()
     await chatPanel.getByText('✨ Context: 14 lines from 2 files').click()
-    await expect(chatPanel.getByRole('button', { name: '@index.html:1-10' })).toBeVisible()
+    await expect(chatPanel.getByRole('button', { name: '@index.html:1-12' })).toBeVisible()
     await expect(
-        chatPanel.getByRole('button', { name: withPlatformSlashes('@lib/batches/env/var.go:1-0') })
+        chatPanel.getByRole('button', { name: withPlatformSlashes('@lib/batches/env/var.go:1-2') })
     ).toBeVisible()
-
-    const expectedEvents = [
-        'CodyVSCodeExtension:menu:command:custom:clicked',
-        'CodyVSCodeExtension:command:custom:executed',
-        'CodyVSCodeExtension:chat-question:submitted',
-        'CodyVSCodeExtension:chat-question:executed',
-    ]
-    await assertEvents(mockServer.loggedEvents, expectedEvents)
 })
 
-test('open and delete cody.json from the custom command menu', async ({ page, sidebar }) => {
+test.extend<ExpectedEvents>({
+    // list of events we expect this test to log, add to this list as needed
+    expectedEvents: [
+        'CodyVSCodeExtension:auth:clickOtherSignInOptions',
+        'CodyVSCodeExtension:login:clicked',
+        'CodyVSCodeExtension:auth:selectSigninMenu',
+        'CodyVSCodeExtension:auth:fromToken',
+        'CodyVSCodeExtension:Auth:connected',
+        'CodyVSCodeExtension:menu:command:custom:clicked',
+        'CodyVSCodeExtension:menu:command:config:clicked',
+    ],
+})('open and delete cody.json from the custom command menu', async ({ page, sidebar }) => {
     // Sign into Cody
     await sidebarSignin(page, sidebar)
 
@@ -240,10 +268,4 @@ test('open and delete cody.json from the custom command menu', async ({ page, si
 
     // The opened cody.json file should be shown as "Deleted"
     await expect(page.getByRole('list').getByLabel(/cody.json(.*)Deleted$/)).toBeVisible()
-
-    const expectedEvents = [
-        'CodyVSCodeExtension:menu:command:custom:clicked',
-        'CodyVSCodeExtension:menu:command:config:clicked',
-    ]
-    await assertEvents(mockServer.loggedEvents, expectedEvents)
 })